Extract shared import-graph.ts: deduplicate graph infrastructure
listProductionSources, extractImportSpecifiers, resolveFileCandidate,
scanSourceFiles, and buildImportGraph were copy-pasted across
hierarchical, behavioral, and hypergraph tests. Extracted into a new
import-graph.ts companion to discover-packages.ts. Each test file now
composes the shared module; ~80 lines of duplicated infrastructure
removed per file.
